DRY up rspec shared contexts correctly
